# docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: autom8_asana API
  description: REST API for Asana integration via autom8_asana SDK
  version: 0.1.0
paths:
...
  /api/v1/tasks:
    get:
      tags:
      - tasks
      summary: List tasks by project or section
      description: "List tasks by project or section with pagination.\n\nArgs:\n \
        \   project: Project GID to list tasks from.\n    section: Section GID to\
        \ list tasks from.\n    limit: Number of items per page (1-100, default 100).\n\
        \    offset: Pagination cursor from previous response.\n\nReturns:\n    List\
        \ of tasks with pagination metadata.\n\nRaises:\n    400: Neither project\
        \ nor section provided, or both provided."
      operationId: list_tasks_api_v1_tasks_get
      parameters:
      - name: project
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Project GID to list tasks from
          title: Project
        description: Project GID to list tasks from
      - name: section
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Section GID to list tasks from
          title: Section
        description: Section GID to list tasks from
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Number of items per page
          default: 100
          title: Limit
        description: Number of items per page
      - name: offset
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Pagination cursor from previous response
          title: Offset
        description: Pagination cursor from previous response
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_AsanaResource__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - tasks
      summary: Create a new task
      description: "Create a new task.\n\nArgs:\n    body: Task creation parameters.\n\
        \nReturns:\n    Created task data.\n\nRaises:\n    400: Neither projects nor\
        \ workspace provided."
      operationId: create_task_api_v1_tasks_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}:
    get:
      tags:
      - tasks
      summary: Get task by GID
      description: "Get a task by its GID.\n\nArgs:\n    gid: Asana task GID.\n  \
        \  opt_fields: Comma-separated list of fields to include in response.\n\n\
        Returns:\n    Task data with requested fields."
      operationId: get_task_api_v1_tasks__gid__get
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: opt_fields
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Comma-separated list of fields to include
          examples:
          - name,notes,due_on,assignee
          title: Opt Fields
        description: Comma-separated list of fields to include
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - tasks
      summary: Update a task
      description: "Update an existing task.\n\nOnly provided fields are updated;\
        \ omitted fields retain their values.\n\nArgs:\n    gid: Asana task GID.\n\
        \    body: Fields to update.\n\nReturns:\n    Updated task data."
      operationId: update_task_api_v1_tasks__gid__put
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - tasks
      summary: Delete a task
      description: "Delete a task.\n\nArgs:\n    gid: Asana task GID.\n\nReturns:\n\
        \    No content on success."
      operationId: delete_task_api_v1_tasks__gid__delete
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/subtasks:
    get:
      tags:
      - tasks
      summary: List subtasks of a task
      description: "List subtasks of a task with pagination.\n\nArgs:\n    gid: Parent\
        \ task GID.\n    limit: Number of items per page (1-100, default 100).\n \
        \   offset: Pagination cursor from previous response.\n\nReturns:\n    List\
        \ of subtasks with pagination metadata."
      operationId: list_subtasks_api_v1_tasks__gid__subtasks_get
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Number of items per page
          default: 100
          title: Limit
        description: Number of items per page
      - name: offset
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Pagination cursor from previous response
          title: Offset
        description: Pagination cursor from previous response
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_AsanaResource__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/dependents:
    get:
      tags:
      - tasks
      summary: List dependent tasks
      description: "List tasks that depend on this task with pagination.\n\nArgs:\n\
        \    gid: Task GID to get dependents for.\n    limit: Number of items per\
        \ page (1-100, default 100).\n    offset: Pagination cursor from previous\
        \ response.\n\nReturns:\n    List of dependent tasks with pagination metadata."
      operationId: list_dependents_api_v1_tasks__gid__dependents_get
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Number of items per page
          default: 100
          title: Limit
        description: Number of items per page
      - name: offset
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Pagination cursor from previous response
          title: Offset
        description: Pagination cursor from previous response
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_AsanaResource__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/duplicate:
    post:
      tags:
      - tasks
      summary: Duplicate a task
      description: "Duplicate a task.\n\nArgs:\n    gid: GID of task to duplicate.\n\
        \    body: Name for the duplicated task.\n\nReturns:\n    New duplicated task\
        \ data."
      operationId: duplicate_task_api_v1_tasks__gid__duplicate_post
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DuplicateTaskRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/tags:
    post:
      tags:
      - tasks
      summary: Add tag to task
      description: "Add a tag to a task.\n\nArgs:\n    gid: Task GID.\n    body: Tag\
        \ GID to add.\n\nReturns:\n    Updated task data."
      operationId: add_tag_api_v1_tasks__gid__tags_post
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTagRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/tags/{tag_gid}:
    delete:
      tags:
      - tasks
      summary: Remove tag from task
      description: "Remove a tag from a task.\n\nArgs:\n    gid: Task GID.\n    tag_gid:\
        \ Tag GID to remove.\n\nReturns:\n    Updated task data."
      operationId: remove_tag_api_v1_tasks__gid__tags__tag_gid__delete
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: tag_gid
        in: path
        required: true
        schema:
          type: string
          title: Tag Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/section:
    post:
      tags:
      - tasks
      summary: Move task to section
      description: "Move a task to a section within a project.\n\nArgs:\n    gid:\
        \ Task GID.\n    body: Target section and project GIDs.\n\nReturns:\n    Updated\
        \ task data."
      operationId: move_to_section_api_v1_tasks__gid__section_post
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveSectionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{gid}/assignee:
    put:
      tags:
      - tasks
      summary: Set task assignee
      description: "Set or clear the task assignee.\n\nArgs:\n    gid: Task GID.\n\
        \    body: Assignee user GID (null to unassign).\n\nReturns:\n    Updated\
        \ task data."
      operationId: set_assignee_api_v1_tasks__gid__assignee_put
      parameters:
      - name: gid
        in: path
        required: true
        schema:
          type: string
          title: Gid
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAssigneeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_AsanaResource_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateTaskRequest:
      properties:
        name:
          type: string
          minLength: 1
          title: Name
          description: Task name
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
          description: Task description
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
          description: Assignee user GID
        projects:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Projects
          description: Project GIDs to add task to
        due_on:
          anyOf:
          - type: string
            pattern: ^\d{4}-\d{2}-\d{2}$
          - type: 'null'
          title: Due On
          description: Due date (YYYY-MM-DD)
        workspace:
          anyOf:
          - type: string
          - type: 'null'
          title: Workspace
          description: Workspace GID (required if no projects)
      additionalProperties: false
      type: object
      required:
      - name
      title: CreateTaskRequest
      description: "Request body for creating a task.\n\nPer FR-API-TASK-003: Create\
        \ task endpoint.\n\nAttributes:\n    name: Task name (required).\n    notes:\
        \ Task description (optional).\n    assignee: Assignee user GID (optional).\n\
        \    projects: List of project GIDs to add task to (optional).\n    due_on:\
        \ Due date in YYYY-MM-DD format (optional).\n    workspace: Workspace GID\
        \ (required if no projects specified)."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SuccessResponse_list_AsanaResource__:
      properties:
        data:
          items:
            $ref: '#/components/schemas/AsanaResource'
          type: array
          title: Data
          description: Response data payload
        meta:
          $ref: '#/components/schemas/ResponseMeta'
          description: Response metadata
      additionalProperties: false
      type: object
      required:
      - data
      - meta
      title: SuccessResponse[list[AsanaResource]]
...